requirements.txt
pyqt5>=5.15.10
psutil>=5.9.5
pywin32>=306
Pyrebase4>=4.7.1
opencv-python>=4.8.0
screen-brightness-control>=0.21.0
comtypes>=1.2.0
pycaw>=20230407
cryptography>=41.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.4.0
